requirements.txt
# Basic Requirements
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3

# Text Processing
nltk==3.8.1
spacy==3.7.2

# Web Framework (for advanced version)
flask==3.0.0
flask-cors==4.0.0

# AI/ML (optional - for advanced features)
openai==1.3.0
groq==0.4.1
anthropic==0.34.2  # For Claude
google-generativeai==0.8.3  # For Gemini
cohere==4.47  # For Cohere

# Database
psycopg2-binary==2.9.9  # PostgreSQL driver (optional if using SQLite)
sqlalchemy==2.0.23
alembic==1.12.1
# SQLite support is built into Python, no extra package needed!

# Utilities
python-dotenv==1.0.0
pandas==2.1.3

# Production Scalability (for 1000+ resumes/day)
celery==5.3.4  # Async task queue
redis==5.0.1  # Queue broker + cache
flask-limiter==3.5.0  # Rate limiting
prometheus-client==0.19.0  # Metrics/monitoring
python-json-logger==2.0.7  # Structured logging

